aider: Modified the exec call to run in the `this.config.basePath` directory.
# Context:
USER: Modify the exec call to run in the this.config.basePath directory
ASSISTANT: To modify the exec call to run in the `this.config.basePath` directory, we need to modify the `callApi` method in the `ScriptCompletionProvider` class in the `src/providers/scriptCompletion.ts` file.

Here is the change:

```typescript
src/providers/scriptCompletion.ts
<<<<<<< ORIGINAL
  async callApi(prompt: string) {
    return new Promise((resolve, reject) => {
      exec(`${this.scriptPath} "${prompt}"`, (error, stdout, stderr) => {
        if (error) {
          reject(error);
        } else {
          resolve({ output: stripText(stdout.trim()) });
        }
      });
    }) as Promise<ProviderResponse>;
  }
=======
  async callApi(prompt: string) {
    return new Promise((resolve, reject) => {
      exec(`cd ${this.config.basePath} && ${this.scriptPath} "${prompt}"`, (error, stdout, stderr) => {
        if (error) {
          reject(error);
        } else {
          resolve({ output: stripText(stdout.trim()) });
        }
      });
    }) as Promise<ProviderResponse>;
  }
>>>>>>> UPDATED
```

This change prepends a `cd` command to change the current directory to `this.config.basePath` before executing the script. Please note that this assumes `this.config.basePath` is always defined and valid. If it can be undefined or invalid, additional error checking should be added.