aider: Changed the file watcher to use `fs.watchFile` instead of `fs.watch` to watch for changes to a symlink.
# Aider chat conversation:

USER: latest.json is a symlink.  Watch for when this symlink is updated in order to emit the update event.

ASSISTANT: To watch for changes to a symlink, we need to use the `fs.watchFile` method instead of `fs.watch`. The `fs.watchFile` method provides more consistent behavior when watching for changes to a symlink.

Here are the changes needed in the `src/web/server.ts` file:

```typescript
src/web/server.ts
<<<<<<< ORIGINAL
  fs.watch(
    latestJsonPath,
    debounce((event: string) => {
      if (event === 'change') {
        socket.emit('update', readLatestJson());
      }
    }, 250),
  );
=======
  fs.watchFile(
    latestJsonPath,
    debounce((curr, prev) => {
      if (curr.mtime !== prev.mtime) {
        socket.emit('update', readLatestJson());
      }
    }, 250),
  );
>>>>>>> UPDATED
```

In the updated code, we're using `fs.watchFile` instead of `fs.watch`. The callback function now takes two arguments, `curr` and `prev`, which represent the current and previous states of the file. We emit the 'update' event if the modification time of the file has changed.