aider: Reset the "recent-files" select menu in ResultsView when recentFiles is updated.
# Aider chat conversation:

USER: Reset the "recent-files" select menu in ResultsView when recentFiles is updated

ASSISTANT: To reset the "recent-files" select menu in ResultsView when recentFiles is updated, we need to modify the `ResultsView.tsx` file. We will introduce a key prop to the `Select` component that changes every time `recentFiles` is updated. This will cause the `Select` component to unmount and remount, effectively resetting its state.

Here is the change:

```tsx
src/web/client/src/ResultsView.tsx
<<<<<<< ORIGINAL
            {recentFiles && recentFiles.length > 0 && (
              <FormControl sx={{ m: 1, minWidth: 200 }} size="small">
                <InputLabel>Eval</InputLabel>
                <Select className="recent-files" label="Previous runs" defaultValue={recentFiles[0]} onChange={(e: SelectChangeEvent) => onRecentFileSelected(e.target.value)}>
=======
            {recentFiles && recentFiles.length > 0 && (
              <FormControl sx={{ m: 1, minWidth: 200 }} size="small">
                <InputLabel>Eval</InputLabel>
                <Select key={recentFiles.join(',')} className="recent-files" label="Previous runs" defaultValue={recentFiles[0]} onChange={(e: SelectChangeEvent) => onRecentFileSelected(e.target.value)}>
>>>>>>> UPDATED
                  {recentFiles.map((file) => (
                    <MenuItem key={file} value={file}>{filenameToDate(file)}</MenuItem>
                  ))}
                </Select>
              </FormControl>
            )}
```

This change will cause the select menu to reset every time the `recentFiles` prop changes.