aider: Change `options` prop of `Autocomplete` to pass entire `ProviderConfig` object instead of just `id`. Change `value` prop of `Autocomplete` to pass entire `ProviderConfig` object instead of just `id`. Change `onChange` prop of `Autocomplete` to directly pass `newValue` array to `onChange` prop. Modify `renderTags` prop of `Autocomplete` to use `ProviderConfig` objects and display `id` of provider.
# Aider chat conversation:

USER: is it possible for autocomplete to retain the entire object in `defaultProviders`, so that in `onChange` it's passed with the `config` object in addition to `id`?

ASSISTANT: Yes, it is possible. We can modify the `options` prop of the `Autocomplete` component to pass the entire `ProviderConfig` object instead of just the `id`. Then, in the `onChange` handler, we can directly pass the `newValue` array to the `onChange` prop, as it will already be an array of `ProviderConfig` objects.

Here are the changes:

```tsx
src/web/webui/pages/setup/ProviderSelector.tsx
<<<<<<< ORIGINAL
        options={defaultProviders.map((provider) => provider.id || 'Unknown provider')}
        value={providers.map((provider) => provider.id || 'Unknown provider')}
        onChange={(event, newValue) => {
          onChange(newValue.map((id) => ({ id })));
        }}
=======
        options={defaultProviders}
        value={providers}
        onChange={(event, newValue: ProviderConfig[]) => {
          onChange(newValue);
        }}
>>>>>>> UPDATED
```

In the `renderTags` prop, we need to modify the `value` parameter to be an array of `ProviderConfig` objects and change `option` to `provider` to reflect this. We also need to change `label={option}` to `label={provider.id || 'Unknown provider'}` to display the `id` of the provider.

Here are the changes:

```tsx
src/web/webui/pages/setup/ProviderSelector.tsx
<<<<<<< ORIGINAL
        renderTags={(value: string[], getTagProps) =>
          value.map((option: string, index: number) => (
            <Chip variant="outlined" label={option} {...getTagProps({ index })} key={index} />
          ))
        }
=======
        renderTags={(value: ProviderConfig[], getTagProps) =>
          value.map((provider: ProviderConfig, index: number) => (
            <Chip variant="outlined" label={provider.id || 'Unknown provider'} {...getTagProps({ index })} key={index} />
          ))
        }
>>>>>>> UPDATED
```

These changes should allow the `Autocomplete` component to retain the entire `ProviderConfig` object and pass it to the `onChange` handler.