aider: Edit ProviderConfigDialog to dynamically build the form fields based on the keys and values of localConfig.
# Aider chat conversation:

USER: Edit ProviderConfigDialog to dynamically build the form fields based on the keys and values of localConfig. It should display string, number, and boolean fields.

ASSISTANT: To dynamically build the form fields based on the keys and values of `localConfig`, we can iterate over the keys of `localConfig` and create a `TextField` for each key. We'll use the key as the label and the corresponding value as the value of the `TextField`.

Here are the changes:

```tsx
src/web/webui/pages/setup/ProviderConfigDialog.tsx
<<<<<<< ORIGINAL
      <DialogContent>
        <TextField
          label="Temperature"
          value={localConfig.temperature}
          onChange={(e) =>
            setLocalConfig({ ...localConfig, temperature: parseFloat(e.target.value) })
          }
        />
        <TextField
          label="Max Tokens"
          value={localConfig.max_tokens}
          onChange={(e) =>
            setLocalConfig({ ...localConfig, max_tokens: parseInt(e.target.value, 10) })
          }
        />
      </DialogContent>
=======
      <DialogContent>
        {Object.keys(localConfig).map((key) => {
          const value = localConfig[key];
          let handleChange;

          if (typeof value === 'number') {
            handleChange = (e) =>
              setLocalConfig({ ...localConfig, [key]: parseFloat(e.target.value) });
          } else if (typeof value === 'boolean') {
            handleChange = (e) =>
              setLocalConfig({ ...localConfig, [key]: e.target.value === 'true' });
          } else {
            handleChange = (e) =>
              setLocalConfig({ ...localConfig, [key]: e.target.value });
          }

          return (
            <TextField
              key={key}
              label={key}
              value={value}
              onChange={handleChange}
            />
          );
        })}
      </DialogContent>
>>>>>>> UPDATED
```

This change will create a `TextField` for each key in `localConfig`, with the key used as the label and the corresponding value used as the value of the `TextField`. The `onChange` handler is determined based on the type of the value. If the value is a number, it's parsed as a float. If it's a boolean, it's converted to a boolean based on whether the new value is the string 'true'. Otherwise, the new value is used as is.