aider: Update `ProviderConfigDialog` to be aesthetically pleasing by adding labels and aligning inputs on new lines.
# Aider chat conversation:

USER: update ProviderConfigDialog to be aesthetically pleasing. Each input should be labeled and on a newline. Inputs should be aligned with each other

ASSISTANT: To make the `ProviderConfigDialog` more aesthetically pleasing, we can add some styling to the `TextField` components. We can use the `Box` component from Material-UI to add some margin to each `TextField`. This will ensure that each input is on a new line. We can also use the `fullWidth` prop of the `TextField` component to make sure that all inputs are aligned with each other.

Here are the changes you need to make in the `ProviderConfigDialog.tsx` file:

```tsx
src/web/webui/pages/setup/ProviderConfigDialog.tsx
<<<<<<< ORIGINAL
            <TextField
              key={key}
              label={key}
              value={value}
              onChange={handleChange}
            />
=======
            <Box mb={2}>
              <TextField
                key={key}
                label={key}
                value={value}
                onChange={handleChange}
                fullWidth
              />
            </Box>
>>>>>>> UPDATED
```

This change wraps each `TextField` in a `Box` component with a bottom margin of `2`. It also adds the `fullWidth` prop to the `TextField` component.